Split up a little extension.js
