+++ src/extension.js
@@ -1,106 +1,8 @@
-let titles = {
-  'default': 'ЕТИС',
-  'stu.teach_plan': 'Учебный план',
-  'ebl_stu.ebl_choice': 'Элективы',
-  'stu.tpr': 'Оставить отзыв',
-  'stu.teachers': 'Преподаватели',
-  'stu.sc_portfolio': 'Портфолио',
-  'stu.timetable': 'Расписание',
-  'stu.change_pass_form': 'Изменить пароль',
-  'stu.change_pass': 'Изменить пароль',
-  'stu_email_pkg.change_email': 'Изменить почту',
-  'stu.announce': 'Объявления',
-  'stu.teacher_notes': 'Сообщения',
-  'est_pkg.show_list': 'Обратная связь',
-  'stu.signs': 'Оценки',
-  'stu.absence': 'Пропущенные',
-  'stu.orders': 'Приказы',
-  'stu_jour.group_tt': 'Журнал посещений',
-  'stu.ses': 'Образовательный стандарт',
-  'stu.electr': 'Электронные ресурсы',
-  'stu.about': 'О ресурсе',
-  'cert_pkg.stu_certif': 'Заказ справок',
-  'stu.term_test': 'Анкетирование',
-  'stu.special_est_list': 'Опросы',
-  'stu.fcl_choice': 'Факультативы',
-};
+import { insertMeta } from "./meta";
+import { titles, changeTitle } from "./titles";
+import { setupTheme } from "./theme";
 
-// Set Theme
-let theme = 'auto';
-let prefersColorSchemeMedia = false;
-if (window.matchMedia) {
-  prefersColorSchemeMedia = window.matchMedia('(prefers-color-scheme: dark)');
-}
-
-function setDarkTheme(e) {
-  document.documentElement.setAttribute('theme', e.matches ? 'dark' : 'light');
-}
-
-function setSystemThemeDetection() {
-  if (window.matchMedia) {
-    prefersColorSchemeMedia = window.matchMedia('(prefers-color-scheme: dark)');
-    document.documentElement.setAttribute('theme', prefersColorSchemeMedia.matches ? 'dark' : 'light');
-    prefersColorSchemeMedia.addEventListener('change', setDarkTheme);
-  }
-}
-
-function removeSystemThemeDetection() {
-  if (window.matchMedia) {
-    prefersColorSchemeMedia.removeEventListener('change', setDarkTheme);
-  }
-}
-
-function detectTheme() {
-  if (localStorage.getItem('theme')) {
-    theme = localStorage.getItem('theme');
-    if (theme == 'auto') {
-      setSystemThemeDetection();
-    } else {
-      document.documentElement.setAttribute('theme', theme);
-    }
-  } else {
-    theme = 'auto';
-    setSystemThemeDetection()
-  }
-}
-
-function switchTheme(e) {
-  if (theme == 'auto') {
-    theme = 'light';
-    e.srcElement.innerHTML = e.srcElement.innerHTML.replace('Системная', 'Светлая');
-    document.documentElement.setAttribute('theme', theme);
-    removeSystemThemeDetection();
-  } else if (theme == 'light') {
-    theme = 'dark';
-    e.srcElement.innerHTML = e.srcElement.innerHTML.replace('Светлая', 'Темная');
-    document.documentElement.setAttribute('theme', theme);
-  } else if (theme == 'dark') {
-    theme = 'auto';
-    e.srcElement.innerHTML = e.srcElement.innerHTML.replace('Темная', 'Системная');
-    setSystemThemeDetection();
-  }
-
-  localStorage.setItem('theme', theme);
-}
-
-detectTheme();
-
-function insertMeta() {
-  const meta = document.createElement('meta');
-  meta.name = 'viewport';
-  meta.content = 'width=device-width, initial-scale=1';
-  document.getElementsByTagName('head')[0].appendChild(meta);
-}
-
-function changeTitle() {
-  let text = titles.default;
-  const page = window.location.pathname.split('/').pop();
-  if (page in titles) {
-    text += ' - ' + titles[page];
-  }
-  const title = document.querySelector('title');
-  title.textContent = text;
-}
+setupTheme();
 
 // Style
 document.addEventListener("DOMContentLoaded", function (event) {
