+++ src/meta.js
@@ -1,0 +1,6 @@
+export function insertMeta() {
+  const meta = document.createElement('meta');
+  meta.name = 'viewport';
+  meta.content = 'width=device-width, initial-scale=1';
+  document.getElementsByTagName('head')[0].appendChild(meta);
+}
